hnsw: narrow vertex.level to uint16
HNSW levels are generated as uint8 and written to the WAL as uint16, so
a 64-bit int per node is wasted space. A lvl() getter keeps the graph
code working with int.

+++ adapters/repos/db/vector/hnsw/commit_logger.go
@@ -351,7 +351,7 @@ func (l *hnswCommitLogger) AddNode(node *vertex) error {
 	l.Lock()
 	defer l.Unlock()
 
-	return l.walWriter.WriteAddNode(node.id, uint16(node.level))
+	return l.walWriter.WriteAddNode(node.id, node.level)
 }
 
 func (l *hnswCommitLogger) SetEntryPointWithMaxLayer(id uint64, level int) error {
